+++ src/test/java/domain/car/CarTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.car;
 
+import domain.random.RangeRandomValueGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
